test: move metrics decoration tests to separate file
